Move help functions to another file
